rename ymdk_np21 to np21; move to ymdk vendor folder
