+++ keyboards/ymdk/np21/np21.c
@@ -16,4 +16,4 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "ymdk_np21.h"
+#include "np21.h"
